Updated Env Definition
- Updated observation space. Increased observation of the state
including target pose (7,); kinematic chain (12,) or (9,) for digit I;
velocity (3,); and contacts with other fingers and the palm (6,)
- Target is now specified as position and orientation of fingertip
- Reward function now includes penalty terms for the angular
displacement and contact with other fingers
- Action space is still the continuous torque value applied to each
tendon
- Distance parameters passed into the environment and returned from the
environment are now all specified in cm.
- Environment modified in such a way that any combination of the 5
digits can be used
- Start states file now needs to be saved in the start_states sub-folder
of the darm_gym_env directory, and passed as a parameter when creating
the environment. This alows for the dynamics of combining different
digits.